# Conflicts:
#   src/data/usecases/waypoint/delete-waypoint.js
#   src/http/controllers/path/delete-path.js
#   src/http/controllers/path/find-path.js
#   src/http/controllers/path/update-path.js

+++ src/data/usecases/map/delete-map.js
@@ -1,9 +1,15 @@
+import { NotFoundError, ServerError } from "../../../http/errors.js";
 import { Maps } from "../../models/map.js";
 
 export async function deleteMap(id) {
+  const map = await Maps.findById(id);
+  if (!map) {
+    throw new NotFoundError(`Map with id ${id} not found`);
+  }
+
   try {
-    return await Maps.findByIdAndDelete(id);
+    await Maps.findByIdAndDelete(id);
   } catch (error) {
-    throw new Error(`Error deleting map: ${error.message}`);
+    throw new ServerError("Error deleting map");
   }
 }
